+++ Controller/Adminhtml/ProductLabel/Index.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Smile\ProductLabel\Controller\Adminhtml\ProductLabel;
 
+use Magento\Backend\Model\View\Result\Page;
 use Magento\Framework\App\Action\HttpGetActionInterface;
 use Magento\Framework\Controller\ResultFactory;
 
@@ -17,13 +18,11 @@ class Index extends AbstractAction implements HttpGetActionInterface
      */
     public function execute()
     {
-        $breadMain = (string) __('Manage Product labels');
-
-        /** @var \Magento\Backend\Model\View\Result\Page $resultPage */
+        /** @var Page $resultPage */
         $resultPage = $this->resultFactory->create(ResultFactory::TYPE_PAGE);
         $resultPage->setActiveMenu('Smile_ProductLabel::manage');
         $resultPage->addBreadcrumb((string) __('Smile Product Label'), (string) __('Smile Product Label'));
-        $resultPage->getConfig()->getTitle()->prepend($breadMain);
+        $resultPage->getConfig()->getTitle()->prepend((string) __('Manage Product labels'));
 
         return $resultPage;
     }