+++ Controller/Adminhtml/ProductLabel/MassStatus.php
@@ -7,6 +7,7 @@ namespace Smile\ProductLabel\Controller\Adminhtml\ProductLabel;
 use Exception;
 use Magento\Framework\App\Action\HttpPostActionInterface;
 use Magento\Framework\Controller\Result\Redirect;
+use Magento\Framework\Exception\LocalizedException;
 use Smile\ProductLabel\Api\Data\ProductLabelInterface;
 
 /**
@@ -17,25 +18,24 @@ class MassStatus extends AbstractAction implements HttpPostActionInterface
     /**
      * Execute action
      *
-     * @throws \Magento\Framework\Exception\LocalizedException|Exception
+     * @throws LocalizedException|Exception
      */
     public function execute(): Redirect
     {
         $collection     = $this->filter->getCollection($this->collectionFactory->create());
         $collectionSize = $collection->getSize();
         $status         = (int) $this->getRequest()->getParam('is_active');
-        // @codingStandardsIgnoreStart
-        $message        = ($status === 0) ? 'A total of %1 product label(s) have been disabled.' : 'A total of %1 product label(s) have been enabled.';
-        // @codingStandardsIgnoreEnd
+        $message        = $status === 0
+            ? 'A total of %1 product label(s) have been disabled.'
+            : 'A total of %1 product label(s) have been enabled.';
 
-        /** @var ProductLabelInterface $plabel */
-        foreach ($collection as $plabel) {
-            $plabel->setIsActive((bool) $status);
-            $this->modelRepository->save($plabel);
+        /** @var ProductLabelInterface $productLabel */
+        foreach ($collection as $productLabel) {
+            $productLabel->setIsActive((bool) $status);
+            $this->modelRepository->save($productLabel);
         }
 
         $this->messageManager->addSuccessMessage((string) __($message, $collectionSize));
-
         $resultRedirect = $this->resultRedirectFactory->create();
 
         return $resultRedirect->setPath('*/*/');