+++ Controller/Adminhtml/ProductLabel/Save.php
@@ -4,10 +4,12 @@ declare(strict_types=1);
 
 namespace Smile\ProductLabel\Controller\Adminhtml\ProductLabel;
 
+use Exception;
 use Magento\Backend\App\Action\Context;
 use Magento\Backend\Model\View\Result\Redirect as ResultRedirect;
 use Magento\Framework\App\Action\HttpPostActionInterface;
 use Magento\Framework\App\Request\DataPersistorInterface;
+use Magento\Framework\App\Request\Http;
 use Magento\Framework\Controller\ResultFactory;
 use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Registry;
@@ -24,17 +26,6 @@ class Save extends AbstractAction implements HttpPostActionInterface
 {
     protected DataPersistorInterface $dataPersistor;
 
-    /**
-     * Save constructor.
-     *
-     * @param Context $context UI Component context
-     * @param Registry $coreRegistry Core Registry
-     * @param ProductLabelFactory $modelFactory Product Label Factory
-     * @param ProductLabelRepository $modelRepository Product Label Repository
-     * @param DataPersistorInterface $dataPersistor Data Persistor
-     * @param Filter $filter Action Filter
-     * @param CollectionFactory $collectionFactory Product Label Collection Factory
-     */
     public function __construct(
         Context $context,
         Registry $coreRegistry,
@@ -45,7 +36,6 @@ class Save extends AbstractAction implements HttpPostActionInterface
         CollectionFactory $collectionFactory
     ) {
         parent::__construct($context, $coreRegistry, $modelFactory, $modelRepository, $filter, $collectionFactory);
-
         $this->dataPersistor = $dataPersistor;
     }
 
@@ -58,7 +48,7 @@ class Save extends AbstractAction implements HttpPostActionInterface
         $resultRedirect = $this->resultFactory->create(ResultFactory::TYPE_REDIRECT);
         $resultRedirect->setPath('*/*/');
 
-        /** @var \Magento\Framework\App\Request\Http $request */
+        /** @var Http $request */
         $request = $this->getRequest();
         $data = $request->getPostValue();
         if (empty($data)) {
@@ -66,16 +56,16 @@ class Save extends AbstractAction implements HttpPostActionInterface
         }
 
         // Get the product label id (if edit).
-        $productlabelId = null;
+        $productLabelId = null;
         if (!empty($data['product_label_id'])) {
-            $productlabelId = (int) $data['product_label_id'];
+            $productLabelId = (int) $data['product_label_id'];
         }
 
         // Load the product label.
-        $model = $this->initModel($productlabelId);
+        $model = $this->initModel($productLabelId);
 
         // By default, redirect to the edit page of the product label.
-        $resultRedirect->setPath('*/*/edit', ['product_label_id' => $productlabelId]);
+        $resultRedirect->setPath('*/*/edit', ['product_label_id' => $productLabelId]);
 
         /** @var ProductLabel $model */
         $model->populateFromArray($data);
@@ -83,7 +73,7 @@ class Save extends AbstractAction implements HttpPostActionInterface
         // Try to save it.
         try {
             $this->modelRepository->save($model);
-            if ($productlabelId === null) {
+            if ($productLabelId === null) {
                 $resultRedirect->setPath('*/*/edit', ['product_label_id' => $model->getProductLabelId()]);
             }
 
@@ -98,7 +88,7 @@ class Save extends AbstractAction implements HttpPostActionInterface
         } catch (LocalizedException $e) {
             $this->messageManager->addErrorMessage($e->getMessage());
             $this->dataPersistor->set('smile_productlabel', $data);
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             $this->messageManager->addExceptionMessage(
                 $e,
                 (string) __('Something went wrong while saving the product label. "%1"', $e->getMessage())