+++ Model/ResourceModel/ProductLabel/Store/ReadHandler.php
@@ -4,17 +4,20 @@ declare(strict_types=1);
 
 namespace Smile\ProductLabel\Model\ResourceModel\ProductLabel\Store;
 
+use Magento\Framework\EntityManager\Operation\ExtensionInterface;
+use Smile\ProductLabel\Model\ResourceModel\ProductLabel;
+
 /**
  * Virtual Attribute store relation read handler.
  */
-class ReadHandler implements \Magento\Framework\EntityManager\Operation\ExtensionInterface
+class ReadHandler implements ExtensionInterface
 {
     /**
      * @inheritdoc
      */
     public function execute($entity, $arguments = [])
     {
-        /** @var \Smile\ProductLabel\Model\ResourceModel\ProductLabel $resource */
+        /** @var ProductLabel $resource */
         $resource = $entity->getResource();
 
         $stores = $resource->getStoreIds($entity);