+++ Model/ResourceModel/ProductLabel/Store/SaveHandler.php
@@ -4,19 +4,21 @@ declare(strict_types=1);
 
 namespace Smile\ProductLabel\Model\ResourceModel\ProductLabel\Store;
 
+use Magento\Framework\EntityManager\Operation\ExtensionInterface;
+use Smile\ProductLabel\Model\ResourceModel\ProductLabel;
+
 /**
  * Virtual Attribute store relation read handler.
  */
-class SaveHandler implements \Magento\Framework\EntityManager\Operation\ExtensionInterface
+class SaveHandler implements ExtensionInterface
 {
     /**
      * @inheritdoc
      */
     public function execute($entity, $arguments = [])
     {
-        /** @var \Smile\ProductLabel\Model\ResourceModel\ProductLabel $resource */
+        /** @var ProductLabel $resource */
         $resource = $entity->getResource();
-
         $resource->saveStoreRelation($entity);
 
         return $entity;