+++ Setup/Patch/Data/UpdateExistingLabelsStoreView.php
@@ -17,25 +17,21 @@ class UpdateExistingLabelsStoreView implements DataPatchInterface
 {
     private ModuleDataSetupInterface $moduleDataSetup;
 
-    /**
-     * @param ModuleDataSetupInterface $moduleDataSetup Module Data Setup
-     */
-    public function __construct(
-        ModuleDataSetupInterface $moduleDataSetup
-    ) {
+    public function __construct(ModuleDataSetupInterface $moduleDataSetup)
+    {
         $this->moduleDataSetup = $moduleDataSetup;
     }
 
     /**
      * @inheritdoc
      */
-    public function apply()
+    public function apply(): self
     {
         $this->moduleDataSetup->getConnection()->startSetup();
 
         $connection = $this->moduleDataSetup->getConnection();
         $select     = $connection->select()->from(
-            ProductLabelInterface::TABLE_NAME,
+            $connection->getTableName(ProductLabelInterface::TABLE_NAME),
             [
                 ProductLabelInterface::PRODUCTLABEL_ID => ProductLabelInterface::PRODUCTLABEL_ID,
                 new Zend_Db_Expr(Store::DEFAULT_STORE_ID . ' as ' . ProductLabelInterface::STORE_ID),
@@ -44,7 +40,7 @@ class UpdateExistingLabelsStoreView implements DataPatchInterface
 
         $data = $connection->fetchAll($select);
         $connection->insertOnDuplicate(
-            ProductLabelInterface::STORE_TABLE_NAME,
+            $connection->getTableName(ProductLabelInterface::STORE_TABLE_NAME),
             $data,
             [ProductLabelInterface::PRODUCTLABEL_ID, ProductLabelInterface::STORE_ID]
         );
@@ -57,7 +53,7 @@ class UpdateExistingLabelsStoreView implements DataPatchInterface
     /**
      * @inheritdoc
      */
-    public static function getDependencies()
+    public static function getDependencies(): array
     {
         return [];
     }
@@ -65,7 +61,7 @@ class UpdateExistingLabelsStoreView implements DataPatchInterface
     /**
      * @inheritdoc
      */
-    public function getAliases()
+    public function getAliases(): array
     {
         return [];
     }